refactor: the output option is unnecessary in molecule cli
